Progress changes
All episodes and arcs are now visible to the users, released or not. Hidden arcs and episodes are still only visible to admins.

I restyled the cards so that the ones in progress are now blue. The hidden ones are now grey and only visible to admins.

I also changed the sorting so that cards and arcs that are in progress are prioritized and pushed to the left top.

I changed the chapter the label in the cards to include the "Chapter" prefix.

I made the text color for the dark grey admin-only cards white for better contrast.

I changed it so that if you input a part that is less than 1 or empty, it's saved as null (Mainly for the Specials arc).

Cards and lists will now update correctly during login/logout.